tijiao

+++ zww-admin/src/main/java/com/stylefeng/guns/common/persistence/dao/TDollOrderItemMapper.java
@@ -19,7 +19,7 @@ import java.util.Map;
  */
 public interface TDollOrderItemMapper extends BaseMapper<TDollOrderItem> {
 
-    List<Map<String, Object>> selectLists(@Param("page") Page<TDollOrderItem> page, @Param("orderId") Integer orderId);
+    List<Map<String, Object>> selectLists(@Param("page") Page<TDollOrderItem> page, @Param("orderId") Long orderId);
 
     //按房间订单id查询
     TDollOrderItem selectByOrderId(Integer orderId);
